+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Zowe CLI package will be documented in this file.
 ## Recent Changes
 
 - Fixed an issue where the CLI exited with status code 0 in case of an error.
+- Added new method "dataSetLike(session, dataSetName, options)" to `Create` class, for use when creating a dataset with parameters like another data set
 
 ## `6.19.1`